+++ src/main/java/com/example/googledriveclone/controller/MinioController.java
@@ -1,36 +1,29 @@
 package com.example.googledriveclone.controller;
 
 import com.example.googledriveclone.services.MinioService;
-import io.minio.BucketExistsArgs;
-import io.minio.ListObjectsArgs;
-import io.minio.MinioClient;
-import io.minio.Result;
 import io.minio.errors.*;
-import io.minio.messages.Bucket;
-import io.minio.messages.Item;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.ModelAttribute;
 
 import java.io.IOException;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
-import java.util.List;
 
-@RestController
+@Controller
 public class MinioController {
 
     @Autowired
     private MinioService minioService;
 
     @GetMapping("/files")
-    public String testMinio() throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException {
+    public String testMinio(Model model) throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException {
 
-        minioService.isBucketExist();
-        minioService.bucketList();
-        minioService.objectList();
-        return "test";
+        model.addAttribute("files", minioService.objectList());
+
+        return "files";
     }
 
 }
