+++ src/main/java/com/example/googledriveclone/services/MinioService.java
@@ -1,51 +1,18 @@
 package com.example.googledriveclone.services;
 
-import io.minio.BucketExistsArgs;
-import io.minio.ListObjectsArgs;
-import io.minio.MinioClient;
 import io.minio.Result;
 import io.minio.errors.*;
-import io.minio.messages.Bucket;
 import io.minio.messages.Item;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-
 import java.io.IOException;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
-import java.util.List;
 
-@Service
-public class MinioService {
-    @Value("${spring.minio.bucket}")
-    private String bucket;
 
-    @Autowired
-    private MinioClient minioClient;
+public interface MinioService {
 
-    public boolean isBucketExist() throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException {
-               return minioClient.bucketExists(BucketExistsArgs.builder().bucket(bucket).build());
-    }
+    boolean isBucketExist() throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException ;
 
-    public  List<Bucket> bucketList() throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException {
-        List<Bucket> bucketList = minioClient.listBuckets();
-        for (Bucket bucket : bucketList) {
-            System.out.println(bucket.creationDate() + ", " + bucket.name());
-        }
+    Iterable<Result<Item>> objectList() throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException;
 
-        return bucketList;
-    }
-
-    public Iterable<Result<Item>> objectList() throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException {
-        // Lists objects information recursively.
-        Iterable<Result<Item>> results = minioClient.listObjects(
-                ListObjectsArgs.builder().bucket(bucket).recursive(true).build());
-
-        for(var item: results){
-            System.out.println(item.get().objectName());
-        }
-
-        return results;
-    }
+    void createFolder(String folderName) throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException;
 }
