+++ src/main/resources/static/css/styles.css
@@ -10,3 +10,13 @@
 .form-container .row{
     justify-content: center;
 }
+.user-files{
+    display: flex;
+    flex-wrap: wrap;
+    padding: 30px 0;
+}
+.file-item{
+    width: 20%;
+    text-align: center;
+    margin-bottom: 20px;
+}
