+++ src/main/resources/templates/files.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html xmlns="http://www.w3.org/1999/xhtml"
+      xmlns:th="http://www.thymeleaf.org">
+<head th:replace="/template-part/header :: header">
+</head>
+<body>
+<div th:replace="/template-part/navigation-menu :: nav"></div>
+<div class="container user-files">
+    <div th:each="item : ${files}" class="file-item">
+        <svg th:if="${item.get().isDir()}" xmlns="http://www.w3.org/2000/svg" x="0px" y="0px"
+             width="64" height="64"
+             viewBox="0,0,256,256"
+             style="fill:#000000;">
+            <g fill="#737373" fill-rule="nonzero" stroke="none" stroke-width="1" stroke-linecap="butt" stroke-linejoin="miter" stroke-miterlimit="10" stroke-dasharray="" stroke-dashoffset="0" font-family="none" font-weight="none" font-size="none" text-anchor="none" style="mix-blend-mode: normal"><g transform="scale(5.12,5.12)"><path d="M5,4c-1.64557,0 -3,1.35554 -3,3v9v10v17c0,1.64446 1.35443,3 3,3h40c1.64506,0 3,-1.35494 3,-3v-17v-10v-5c0,-1.64506 -1.35494,-3 -3,-3h-27c0.08657,0 -0.03101,0.00036 -0.27539,-0.28125c-0.24438,-0.28161 -0.54519,-0.74881 -0.85937,-1.25c-0.31418,-0.50119 -0.64346,-1.03596 -1.05859,-1.50586c-0.41514,-0.4699 -0.98789,-0.96289 -1.80664,-0.96289zM5,6h9c-0.06075,0 0.06114,0.00701 0.30859,0.28711c0.24746,0.2801 0.54864,0.74533 0.86133,1.24414c0.31269,0.49881 0.63651,1.03161 1.04297,1.5c0.40646,0.46839 0.96604,0.96875 1.78711,0.96875h27c0.56294,0 1,0.43706 1,1v2.1875c-0.31489,-0.11356 -0.64816,-0.1875 -1,-0.1875h-40c-0.35184,0 -0.68511,0.07394 -1,0.1875v-6.1875c0,-0.56354 0.43557,-1 1,-1zM5,15h40c0.56503,0 1,0.43497 1,1v10v17c0,0.56294 -0.43706,1 -1,1h-40c-0.56443,0 -1,-0.43646 -1,-1v-17v-10c0,-0.56503 0.43497,-1 1,-1z"></path></g></g>
+        </svg>
+        <svg th:unless="${item.get().isDir()}" xmlns="http://www.w3.org/2000/svg" x="0px" y="0px"
+             width="64" height="64"
+             viewBox="0,0,256,256"
+             style="fill:#000000;">
+            <g fill="#737373" fill-rule="nonzero" stroke="none" stroke-width="1" stroke-linecap="butt" stroke-linejoin="miter" stroke-miterlimit="10" stroke-dasharray="" stroke-dashoffset="0" font-family="none" font-weight="none" font-size="none" text-anchor="none" style="mix-blend-mode: normal"><g transform="scale(4,4)"><path d="M18.50586,9.41406c-2.92848,0 -5.32422,2.39574 -5.32422,5.32422v35.41602c0,2.92848 2.39574,5.32422 5.32422,5.32422h28.13867c2.92848,0 5.32422,-2.39574 5.32422,-5.32422v-21.00586c0.0051,-0.36064 -0.18438,-0.69608 -0.49587,-0.87789c-0.3115,-0.18181 -0.69676,-0.18181 -1.00825,0c-0.3115,0.18181 -0.50097,0.51725 -0.49587,0.87789v21.00586c0,1.84752 -1.4767,3.32422 -3.32422,3.32422h-28.13867c-1.84752,0 -3.32422,-1.4767 -3.32422,-3.32422v-35.41602c0,-1.84752 1.4767,-3.32422 3.32422,-3.32422h18.20703l11.73633,12.61719h-9.38477c-0.7814,0 -1.39453,-0.61117 -1.39453,-1.39258v-5.83984c0.0051,-0.36064 -0.18438,-0.69608 -0.49587,-0.87789c-0.3115,-0.18181 -0.69676,-0.18181 -1.00825,0c-0.3115,0.18181 -0.50097,0.51725 -0.49587,0.87789v5.83984c0,1.86259 1.53194,3.39258 3.39453,3.39258h11.67969c0.39731,-0.00052 0.75663,-0.23617 0.91543,-0.60036c0.1588,-0.36419 0.08697,-0.78784 -0.18301,-1.07933l-13.5957,-14.61719c-0.189,-0.20394 -0.45437,-0.32 -0.73242,-0.32031z"></path></g></g>
+        </svg>
+        <p th:text="${item.get().objectName()}"></p>
+    </div>
+</div>
+<div th:replace="/template-part/footer :: footer"></div>
+</body>
+</html>
